requirements.txt
fastapi
uvicorn
requests
python-dotenv
qdrant-client
cohere